pyproject.toml
[tool.poetry]
name = "desdeo-website"
version = "0.1.0"
description = "The website for the DESDEO framework implemented using Django."
authors = ["Giovanni Misitano <giovanni.a.misitano@jyu.fi>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
flake8 = "^3.7"
black = "^19.3b0"
Pillow = "^6.2"
Markdown = "^3.1"
Markdown2 = "^2.3"
django-stdimage = "^5.0"
django-cleanup = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
